task adding functionality works
